Move some files to new service-update-flow folder
