+++ libs/domains/services/feature/src/lib/service-blueprint-update-flow/blueprint-update-context.tsx
@@ -5,7 +5,7 @@ import {
 } from 'qovery-typescript-axios'
 import { type ReactNode, createContext, useContext } from 'react'
 import { type AnyService } from '@qovery/domains/services/data-access'
-import { type BlueprintFieldValue, type BlueprintFieldValues } from '../../blueprint-field-utils/blueprint-field-utils'
+import { type BlueprintFieldValue, type BlueprintFieldValues } from '../blueprint-field-utils/blueprint-field-utils'
 import {
   type BlueprintUpdateEditableValue,
   type BlueprintUpdateReviewSection,
