+++ libs/domains/services/feature/src/lib/service-blueprint-update-flow/blueprint-update-flow.tsx
@@ -6,11 +6,11 @@ import {
   type BlueprintFieldValue,
   type BlueprintFieldValues,
   isFieldValid,
-} from '../../blueprint-field-utils/blueprint-field-utils'
-import { useBlueprintUpdate } from '../../hooks/use-blueprint-update/use-blueprint-update'
-import { useDeployBlueprint } from '../../hooks/use-deploy-blueprint/use-deploy-blueprint'
-import { usePreviewBlueprintUpdate } from '../../hooks/use-preview-blueprint-update/use-preview-blueprint-update'
-import { useUpdateBlueprint } from '../../hooks/use-update-blueprint/use-update-blueprint'
+} from '../blueprint-field-utils/blueprint-field-utils'
+import { useBlueprintUpdate } from '../hooks/use-blueprint-update/use-blueprint-update'
+import { useDeployBlueprint } from '../hooks/use-deploy-blueprint/use-deploy-blueprint'
+import { usePreviewBlueprintUpdate } from '../hooks/use-preview-blueprint-update/use-preview-blueprint-update'
+import { useUpdateBlueprint } from '../hooks/use-update-blueprint/use-update-blueprint'
 import { BlueprintUpdateFlowProvider } from './blueprint-update-context'
 import { BlueprintUpdateFlowShell } from './blueprint-update-flow-shell'
 import {
